Fix Integration-tests bug

+++ integration-tests/src/main/java/io/quarkiverse/jimmer/it/service/impl/BookImpl.java
@@ -25,7 +25,7 @@ public class BookImpl implements IBook {
     BookTable table = BOOK_TABLE;
 
     @Override
-    public Book findById(int id) {
+    public Book findById(long id) {
         return Jimmer.getDefaultJSqlClient().findById(Book.class, id);
     }
